Tweak: Changing the use-upload-file hook into use-ajax hook with a generic and reusable structure.
